[wacoq] Modularize the implementation.
In particular, this isolates the package manager a bit better, in the
hopes of moving it to a callback.
