+++ backend/wasm/core.ts
@@ -1,10 +1,13 @@
 import { EventEmitter } from 'events';
-import { unzipSync } from 'fflate';
 import { isBrowser, isWebWorker } from 'browser-or-node';
 import { OCamlExecutable, OCamlCAPI } from './ocaml_exec';
 import { WorkerInterrupts } from './interrupt';
+import { IO, FetchMode } from './io';
+import { PackageManager } from './pkgs';
 
-class IcoqPod extends EventEmitter {
+const DEFAULT_FETCH_MODE: FetchMode = (isBrowser || isWebWorker) ? 'browser' : 'fs';
+
+export class IcoqPod extends EventEmitter {
 
     core: OCamlExecutable
     intr: WorkerInterrupts
@@ -12,6 +15,7 @@ class IcoqPod extends EventEmitter {
     binDir: string
     nmDir: string
     io: IO
+    pkgs: PackageManager
 
     constructor(binDir?: string, nmDir?: string, fetchMode: FetchMode = DEFAULT_FETCH_MODE) {
         super();
@@ -26,12 +30,8 @@ class IcoqPod extends EventEmitter {
 
         this.io = new IO(fetchMode);
         this.intr = new WorkerInterrupts();
-    }
 
-    async findlibStartup() {
-        this.putFile('/lib/findlib.conf', `path="/lib/ocaml"`);
-
-        // await this.unzip('/scratch/lib.zip', '/lib/ocaml');
+        this.pkgs = new PackageManager(this.emit, this.answer, this.putFile, this.io, this.binDir);
     }
 
     get fs() { return this.core.fs; }
@@ -50,50 +50,6 @@ class IcoqPod extends EventEmitter {
         this.putFile(toPath, content);
     }
 
-    loadPackage(uri: string, refresh: boolean = true) {
-        return this.loadPackages([uri], refresh);
-    }
-
-    async loadPackages(uris: string | string[], refresh: boolean = true) {
-        if (typeof uris == 'string') uris = [uris];
-
-        await Promise.all(uris.map(async uri => {
-            try {
-                await this.unzip(uri, '/lib');
-                this.answer([['LibLoaded', uri]]);
-            }
-            catch (e) {
-                this.answer([['LibError', uri, e.toString()]]);
-                throw e;
-            }
-        }));
-
-        //if (refresh)
-        //    this.command(['RefreshLoadPath']);
-
-        this.answer([['LoadedPkg', uris]]);
-    }
-
-    async loadSources(uri: string, dirpath: string) {
-        var subdir = dirpath.replace(/[.]|([^/])$/g, '$1/');
-        this.unzip(uri, `/src/${subdir}`);
-    }
-
-    unzip(uri: string, dir: string) {
-        return this.io.unzip(this._pkgUri(uri),
-                    (fn, ui8a) => this.putFile(`${dir}/${fn}`, ui8a),
-                    p => this._progress(uri, p));
-    }
-
-    _pkgUri(uri: string) {
-        return (uri[0] == '+') ?
-            `${this.binDir}/coq/${uri.substring(1)}.coq-pkg` : uri;
-    }
-
-    _progress(uri: string, download: DownloadProgress) {
-        this.emit('progress', {uri, download});
-    }
-
     putFile(filename: string, content: Uint8Array | string) {
         if (!filename.startsWith('/')) filename = `/lib/${filename}`;
         // needs to be synchronous
@@ -110,7 +66,7 @@ class IcoqPod extends EventEmitter {
 
     command(cmd: any[]) {
         switch (cmd[0]) {
-        case 'LoadPkg':        this.loadPackages(cmd[1]);          return;
+        case 'LoadPkg':        this.pkgs.loadPackages(cmd[1]);     return;
         case 'Put':            this.putFile(cmd[1], cmd[2]);       return;
         case 'Get':            this.getFile(cmd[1]);               return;
         case 'InterruptSetup': this.intr.setup(cmd[1]);            return;
@@ -176,70 +132,3 @@ class IcoqPod extends EventEmitter {
         );
     }
 }
-
-
-class IO {
-
-    pending = new Set<ReadableStreamDefaultReader<Uint8Array>>()
-
-    constructor(public fetchMode: FetchMode) { }
-
-    async unzip(uri: string, put: (filename: string, content: Uint8Array) => void, progress?: (p: any) => void) {
-        var zip = unzipSync(await this._fetch(uri, progress));
-
-        for (let [filename, data] of Object.entries(zip)) {
-            if (data.length > 0)  /* regular file; assumes no empty files? */
-                put(filename, data);
-        }
-    }
-
-    async _fetch(uri: string, progress?: (p: any) => void) : Promise<Uint8Array> {
-        if (progress && this.fetchMode === 'browser') {
-            return this._toU8A(this._fetchWithProgress(uri, progress));
-        }
-        else return this._fetchSimple(uri);
-    }
-
-    async _toU8A(blob: Promise<Blob>) {
-        return new Uint8Array(await (await blob).arrayBuffer());
-    }
-
-    async _fetchSimple(uri: string) {
-        switch (this.fetchMode) {
-        case 'browser':
-            return new Uint8Array(await (await fetch(uri)).arrayBuffer());
-        case 'fs':
-            const fs = require('fs');
-            return (0||fs.readFileSync)(uri);
-        }
-    }
-
-    // boilerplate
-    async _fetchWithProgress(uri: string, progress: (p: any) => void) {
-        var response = await fetch(uri),
-            total = +response.headers.get('Content-Length') || 1000,
-            r = response.body.getReader(), chunks = [], downloaded = 0;
-        this.pending.add(r); // prevent reader from being disposed
-        try {
-            for(;;) {
-                var {value, done} = await r.read();
-                if (done) break;
-                chunks.push(value);
-                downloaded += value.length;
-                progress({total, downloaded})
-            }
-            return new Blob(chunks);
-        }
-        finally { this.pending.delete(r); }
-    }
-
-}
-
-
-type FetchMode = 'browser' | 'fs';
-type DownloadProgress = { total: number, downloaded: number };
-
-const DEFAULT_FETCH_MODE: FetchMode = (isBrowser || isWebWorker) ? 'browser' : 'fs';
-
-
-export  { IcoqPod, DownloadProgress }
